requirements.txt
colorama==0.3.9
decorator==4.2.1
gitdb2==2.0.2
GitPython==2.1.5
ipython==6.2.1
ipython-genutils==0.2.0
javalang==0.11.0
jedi==0.11.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
PyGithub==1.38
Pygments==2.2.0
pygount==1.0.0
PyJWT==1.6.1
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.3
traitlets==4.3.2
wcwidth==0.1.7
